# [`Optimize`]: Cache implemetation