--- a/src/Neo.IO/Caching/Cache.cs
+++ b/src/Neo.IO/Caching/Cache.cs
@@ -13,6 +13,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Threading;
 
 namespace Neo.IO.Caching
@@ -20,28 +21,79 @@ namespace Neo.IO.Caching
     public abstract class Cache<TKey, TValue>(int maxCapacity, IEqualityComparer<TKey>? comparer = null)
         : ICollection<TValue>, IDisposable where TKey : notnull
     {
-        protected record class CacheItem(TKey Key, TValue Value)
+        protected class CacheItem
         {
-            public readonly DateTime Time = DateTime.UtcNow;
+            public readonly TKey Key;
+            public readonly TValue Value;
+
+            private CacheItem _prev, _next;
+
+            public CacheItem(TKey key, TValue value)
+            {
+                Key = key;
+                Value = value;
+                _prev = this;
+                _next = this;
+            }
+
+            public bool IsEmpty => ReferenceEquals(_prev, this);
+
+            /// <summary>
+            /// Adds an item after the current item.
+            /// </summary>
+            /// <param name="another">The item to add.</param>
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            public void Add(CacheItem another) => another.Link(this, _next);
+
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            private void Link(CacheItem prev, CacheItem next)
+            {
+                _prev = prev;
+                _next = next;
+                prev._next = this;
+                next._prev = this;
+            }
+
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            public void Unlink()
+            {
+                _prev._next = _next;
+                _next._prev = _prev;
+                _prev = this;
+                _next = this;
+            }
+
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            public CacheItem? RemovePrevious()
+            {
+                if (IsEmpty) return null;
+                var prev = _prev;
+                prev.Unlink();
+                return prev;
+            }
         }
 
-        protected readonly ReaderWriterLockSlim RwSyncRootLock = new(LockRecursionPolicy.SupportsRecursion);
-        protected readonly Dictionary<TKey, CacheItem> InnerDictionary = new(comparer);
+        // key-value in head is invalid, and it is used as a sentinel.
+        protected readonly CacheItem _head = new(default!, default!);
+
+#if NET9_0_OR_GREATER
+        protected readonly Lock _lock = new();
+#else
+        protected readonly object _lock = new();
+#endif
+
+        protected readonly Dictionary<TKey, CacheItem> _innerDictionary = new(comparer);
 
         public TValue this[TKey key]
         {
             get
             {
-                RwSyncRootLock.EnterReadLock();
-                try
+                lock (_lock)
                 {
-                    if (!InnerDictionary.TryGetValue(key, out var cached)) throw new KeyNotFoundException();
-                    OnAccess(cached);
-                    return cached.Value;
-                }
-                finally
-                {
-                    RwSyncRootLock.ExitReadLock();
+                    if (!_innerDictionary.TryGetValue(key, out var item))
+                        throw new KeyNotFoundException();
+                    OnAccess(item);
+                    return item.Value;
                 }
             }
         }
@@ -50,14 +102,9 @@ namespace Neo.IO.Caching
         {
             get
             {
-                RwSyncRootLock.EnterReadLock();
-                try
+                lock (_lock)
                 {
-                    return InnerDictionary.Count;
-                }
-                finally
-                {
-                    RwSyncRootLock.ExitReadLock();
+                    return _innerDictionary.Count;
                 }
             }
         }
@@ -67,41 +114,35 @@ namespace Neo.IO.Caching
         public void Add(TValue item)
         {
             var key = GetKeyForItem(item);
-            RwSyncRootLock.EnterWriteLock();
-            try
+            lock (_lock)
             {
                 AddInternal(key, item);
-            }
-            finally
-            {
-                RwSyncRootLock.ExitWriteLock();
             }
         }
 
         private void AddInternal(TKey key, TValue item)
         {
-            if (InnerDictionary.TryGetValue(key, out var cached))
+            if (_innerDictionary.TryGetValue(key, out var cached))
             {
                 OnAccess(cached);
             }
             else
             {
-                if (InnerDictionary.Count >= maxCapacity)
+                if (_innerDictionary.Count >= maxCapacity)
                 {
-                    var removedCount = InnerDictionary.Count - maxCapacity + 1;
-                    foreach (var toDelete in InnerDictionary.Values.OrderBy(p => p.Time).Take(removedCount))
-                    {
-                        RemoveInternal(toDelete);
-                    }
+                    var prev = _head.RemovePrevious();
+                    if (prev is not null) RemoveInternal(prev);
                 }
-                InnerDictionary.Add(key, new(key, item));
+
+                var added = new CacheItem(key, item);
+                _innerDictionary.Add(key, added);
+                _head.Add(added);
             }
         }
 
         public void AddRange(IEnumerable<TValue> items)
         {
-            RwSyncRootLock.EnterWriteLock();
-            try
+            lock (_lock)
             {
                 foreach (var item in items)
                 {
@@ -109,40 +150,31 @@ namespace Neo.IO.Caching
                     AddInternal(key, item);
                 }
             }
-            finally
-            {
-                RwSyncRootLock.ExitWriteLock();
-            }
         }
 
         public void Clear()
         {
-            RwSyncRootLock.EnterWriteLock();
-            try
+            CacheItem[] items;
+            lock (_lock)
             {
-                foreach (var toDelete in InnerDictionary.Values.ToArray())
-                {
-                    RemoveInternal(toDelete);
-                }
+                items = _innerDictionary.Values.ToArray();
+                _innerDictionary.Clear();
+                _head.Unlink();
             }
-            finally
+
+            foreach (var item in items)
             {
-                RwSyncRootLock.ExitWriteLock();
+                if (item.Value is IDisposable disposable) disposable.Dispose();
             }
         }
 
         public bool Contains(TKey key)
         {
-            RwSyncRootLock.EnterReadLock();
-            try
+            lock (_lock)
             {
-                if (!InnerDictionary.TryGetValue(key, out var cached)) return false;
+                if (!_innerDictionary.TryGetValue(key, out var cached)) return false;
                 OnAccess(cached);
                 return true;
-            }
-            finally
-            {
-                RwSyncRootLock.ExitReadLock();
             }
         }
 
@@ -156,63 +188,53 @@ namespace Neo.IO.Caching
             if (array == null) throw new ArgumentNullException(nameof(array));
             if (startIndex < 0) throw new ArgumentOutOfRangeException(nameof(startIndex));
 
-            RwSyncRootLock.EnterReadLock();
-            try
+            lock (_lock)
             {
-                if (startIndex + InnerDictionary.Count > array.Length)
-                    throw new ArgumentOutOfRangeException(nameof(startIndex));
-                foreach (var item in InnerDictionary.Values)
+                var count = _innerDictionary.Count;
+                if (startIndex + count > array.Length)
+                {
+                    throw new ArgumentOutOfRangeException(nameof(startIndex),
+                        $"startIndex({startIndex}) + Count({count}) > Length({array.Length})");
+                }
+
+                foreach (var item in _innerDictionary.Values)
                 {
                     array[startIndex++] = item.Value;
                 }
-            }
-            finally
-            {
-                RwSyncRootLock.ExitReadLock();
             }
         }
 
         public void Dispose()
         {
             Clear();
-            RwSyncRootLock.Dispose();
         }
 
         public IEnumerator<TValue> GetEnumerator()
         {
-            RwSyncRootLock.EnterReadLock();
-            try
+            TValue[] values;
+            lock (_lock)
             {
-                foreach (var item in InnerDictionary.Values.Select(p => p.Value))
+                var index = 0;
+                values = new TValue[_innerDictionary.Count];
+                foreach (var item in _innerDictionary.Values)
                 {
-                    yield return item;
+                    values[index++] = item.Value;
                 }
             }
-            finally
-            {
-                RwSyncRootLock.ExitReadLock();
-            }
+            return values.AsEnumerable().GetEnumerator();
         }
 
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
         protected abstract TKey GetKeyForItem(TValue item);
 
         public bool Remove(TKey key)
         {
-            RwSyncRootLock.EnterWriteLock();
-            try
+            lock (_lock)
             {
-                if (!InnerDictionary.TryGetValue(key, out var cached)) return false;
+                if (!_innerDictionary.TryGetValue(key, out var cached)) return false;
                 RemoveInternal(cached);
                 return true;
-            }
-            finally
-            {
-                RwSyncRootLock.ExitWriteLock();
             }
         }
 
@@ -225,7 +247,8 @@ namespace Neo.IO.Caching
 
         private void RemoveInternal(CacheItem item)
         {
-            InnerDictionary.Remove(item.Key);
+            _innerDictionary.Remove(item.Key);
+            item.Unlink();
             if (item.Value is IDisposable disposable)
             {
                 disposable.Dispose();
@@ -234,19 +257,14 @@ namespace Neo.IO.Caching
 
         public bool TryGet(TKey key, out TValue item)
         {
-            RwSyncRootLock.EnterReadLock();
-            try
+            lock (_lock)
             {
-                if (InnerDictionary.TryGetValue(key, out var cached))
+                if (_innerDictionary.TryGetValue(key, out var cached))
                 {
                     OnAccess(cached);
                     item = cached.Value;
                     return true;
                 }
-            }
-            finally
-            {
-                RwSyncRootLock.ExitReadLock();
             }
             item = default!;
             return false;