# Description

Optimize `Cache` implemetation:
1. support FIFO and LRU;
2. better performance;

Before:
```
.NET SDK 9.0.202
  [Host]     : .NET 9.0.3 (9.0.325.11113), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 9.0.3 (9.0.325.11113), Arm64 RyuJIT AdvSIMD


| Method            | OperationCount | Mean      | Error    | StdDev   |
|------------------ |--------------- |----------:|---------:|---------:|
| FIFOCacheAdd      | 1000           | 467.94 us | 1.612 us | 1.507 us |
| FIFOCacheContains | 1000           |  10.09 us | 0.006 us | 0.005 us |
| FIFOCacheAdd      | 10000          | 471.67 us | 2.741 us | 2.564 us |
| FIFOCacheContains | 10000          |  10.13 us | 0.027 us | 0.025 us |
````

After:
```
  [Host]     : .NET 9.0.3 (9.0.325.11113), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 9.0.3 (9.0.325.11113), Arm64 RyuJIT AdvSIMD


| Method            | OperationCount | Mean      | Error     | StdDev    |
|------------------ |--------------- |----------:|----------:|----------:|
| FIFOCacheAdd      | 1000           | 28.503 us | 0.1167 us | 0.0975 us |
| FIFOCacheContains | 1000           |  5.285 us | 0.0087 us | 0.0073 us |
| FIFOCacheAdd      | 10000          | 28.331 us | 0.0773 us | 0.0645 us |
| FIFOCacheContains | 10000          |  5.286 us | 0.0031 us | 0.0029 us |
```

Fixes # (issue)

## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Optimization (the change is only an optimization)
- [ ] Style (the change is only a code style for better maintenance or standard purpose)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update



# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules